requirements.txt
fastapi
uvicorn
jinja2
python-dotenv
cloudinary
psycopg2-binary
requests
pydantic
